pyproject.toml
[tool.poetry]
name = "rchat"
version = "0.3.0"
description = "A simple RocketChat command line client"
authors = ["Javier Santacruz <javier.santacruz.lc@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"
homepage = "https://github.com/jvrsantacruz/rchat"
repository = "https://github.com/jvrsantacruz/rchat"
documentation = "https://github.com/jvrsantacruz/rchat"
keywords = ["rocketchat", "command-line", "python"]
classifiers = [
  "Topic :: Communications :: Chat",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Programming Language :: Python :: 3.7",
]

[tool.poetry.dependencies]
python = "^3.11"
click = "*"
rocketchat-API = "*"
confight = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
pyHamcrest = "*"
flake8 = "*"
mypy = "*"
wheel = "*"
pylint = "*"
types-requests = "^2.31.0.20240125"
pyright = "^1.1.349"

[tool.poetry.scripts]
rchat = "rchat.cli:cli"

[tool.black]
line-length = 79
target-version = ["py37"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
